fix: move ng-package.schema.json to dist root directory
BREAKING CHANGES: the `ng-package.schema.json` was accidentally moved to `lib` folder in a previous release. Restore it in its original location!
