+++ src/ng-package.schema.json
