+++ tsconfig.packagr.json
@@ -8,7 +8,7 @@
     "inlineSources": true,
     "declaration": true,
     "lib": ["es2015", "dom"],
-    "outDir": "dist/lib",
+    "outDir": "dist",
     "experimentalDecorators": true,
     "noEmitOnError": true,
     "noImplicitAny": false,
